Split index file into 2 files
